requirements.txt
blinker
flask
flask-dance
flask-login
flask-sqlalchemy
python-dotenv
pymysql
sqlalchemy
Sqlalchemy-utils
